Add Android Support for Firebase Push Notifications
Also add .editorconfig file

And it works. Including being able to set the `largeIcon` using the
`icon` slot in data messages.

Message format is slight different (and more correct) than in
Fuse.PushNotifications:

An example data message
```
POST https://fcm.googleapis.com/fcm/send
Authorization: :auth-key
Content-Type: application/json

{
	"registration_ids": [:reg-id],
	"data": {
		"title": "Well would ya look at!",
		"body": "Hello from some app",
		"icon": "assets/thing.png"
	}
}
```
Here, if the app is closed or in the background, Fuse will make
the notification in the notification draw. It will use the specified
icon even if in a FuseBundle rather than an android 'resource' (in the
res folder)

An example notification message
```
POST https://fcm.googleapis.com/fcm/send
Authorization: :auth-key
Content-Type: application/json

{
	"registration_ids": [:reg-id],
	"notification" : {
		"body" : "great match!",
		"title" : "Portugal vs. Denmark",
		"icon" : "myicon",
		"click_action": "fuseFirebaseBackgroundNotify"
    }
}
```
Here, if the app is closed or in the background, firebase will make
the notification in the notification draw. It will not be able to use
icons that were included using BundleFiles. The  `click_action` is
mandatory and not currently configurable. It is needed to tell Fuse
which internal system needed the intent.

In both cases if the message arrived whilst the app is in the
foreground then no notification is put in the notifcation draw, the
message is simply delivered to the app.

On the JS side, it is almost identical to Fuse.PushNotifications
except that `registrationSucceeded` may be fired many times during an
app's lifecycle. When this happens is up to Firebase so always have a
listener so you can send the details to your server

No iOS support for now.