Made Android version of Firebase.Database.read compatible with nulls, lists and primitive types
Previously, writing require("Firebase/Database").read(key) would result in a crash if the value being read was either null (non-existent) or a primitive type. Additionally, if a List contained a String of a number, this String would be transformed into a number, instead of the desired String. By rewriting Read for Android in Database.uno, reading a List or primitive type will now result in resolve being called with a stringified version of the List or primitive type, so that calling JSON.parse(...) on the result will give back the exact input value, while reading a null-value results in resolve being called with the value null.

#56