Made iOS version of Firebase.Database.read compatible with booleans and numbers, and cleaned up the iOS code for Read in Database.uno
Previously, writing require("Firebase/Database").read(key) would result in a crash if the value being read was an NSNumber (number or boolean). By rewriting Read for iOS in Database.uno, reading a boolean or number will now result in resolve being called with a stringified NSNumber, so that calling JSON.parse(...) on the result will give back the exact number or a 1 or 0 for a boolean.

#56